Fix successive signals breaking interrupt in tests
Pytest could not be interrupted on Windows, as it ignored ctrl-c.
This was the case for the full pytest execution, as soon as one test sent several termination signals to the same subprocess, which was the case for the few first tests (integration connection tests).
Subprocesses are now polled for exit before sending a signal and signals in the tests themselves are followed by a sleep.

+++ test/integration/context.py
@@ -73,12 +73,15 @@ class Context: # pylint: disable = too-many-instance-attributes
 			database_client = pymongo.MongoClient(self.database_uri)
 			database_client.drop_database(database_client.get_database())
 			database_client.close()
+
 		for process in self.process_collection:
-			os.kill(process.pid, shutdown_signal)
-			try:
-				process.wait(5)
-			except subprocess.TimeoutExpired:
-				process.kill()
+			if process.poll() is None:
+				os.kill(process.pid, shutdown_signal)
+				try:
+					process.wait(5)
+				except subprocess.TimeoutExpired:
+					process.kill()
+
 		self.process_collection.clear()
